rpmrepocloner: Always refresh the tdnf cache
Refreshing the tdnf cache is a workaround for an issue where
the new --targetarch option does not independently refresh the
cache. So if the cache were originally x86_64/noarch and you
attempted to tdnf --targetarch aarch64, then tdnf would return
no package found.

Eventually goal is to fix the tdnf patch and remove this
workaround

Signed-off-by: Chris Co <chrco@microsoft.com>